Continue refactoring testing utilities
Continue refactoring testing utilities into its own crate. This
moves the testing utilities of the following crates:
- consensus
- validator
