+++ test-utils/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod blockchain;
+pub mod consensus;
+pub mod node;
+pub mod validator;
