+++ validator/Cargo.toml
@@ -63,6 +63,7 @@ tokio = { version = "1.14", features = ["rt", "test-util", "time"] }
 nimiq-build-tools = { path = "../build-tools" }
 nimiq-network-libp2p = { path = "../network-libp2p" }
 nimiq-network-mock = { path = "../network-mock" }
+nimiq-test-utils = { path = "../test-utils" }
 
 [features]
 metrics = []
